+++ src/test/fixtures/source-map/inline-map/main.ts
@@ -2,6 +2,7 @@ interface Greeting {
   name: string;
 }
 
+// @ts-ignore
 function hello({name}: Greeting): void {
   console.log(`Hello, ${name}!`);
 }